exported function, returned form submission to log
