+++ test/tests.js
@@ -1,17 +1,5 @@
+import getApplicant from '../src/getApplicant.js';
 const test = QUnit.test;
-function getApplicant(formData) {
-
-    //Make object literal, literally making an object
-    const applicant = {
-        name: formData.get('name'),
-        age: formData.get('age'),
-        appSkills: formData.getAll('app-skills'),
-        found: formData.get('found'),
-        role: formData.get('role'),
-        motivation: formData.get('motivation'),
-    }
-    return applicant;
-}
 
 test('time to test a function', function(assert) {
     //Arrange
@@ -23,9 +11,10 @@ test('time to test a function', function(assert) {
         role: 'helicopter',
         motivation: 'i like caleb'
     };
+
     // Set up your parameters and expectations
     const formData = new FormData();
-    formData.set('name', expected.name);
+    formData.set('full-name', expected.name);
     formData.set('age', expected.age);
     formData.set('app-skills', expected.appSkills[0]);
     formData.append('app-skills', expected.appSkills[1]);
@@ -34,8 +23,6 @@ test('time to test a function', function(assert) {
     formData.set('motivation', expected.motivation);
     //Act 
     const applicant = getApplicant(formData);
-
-    // Call the function you're testing and set the result to a const
 
     //Assert
     assert.deepEqual(applicant, expected);
